+++ AlcedoAndroid/app/src/main/cpp/core/edit/operators/color_wheel_op.cpp
@@ -13,15 +13,15 @@ void ColorWheelOperator::apply_rgb(float* pixels, int width, int height,
                                     float gamma_r, float gamma_g, float gamma_b,
                                     float gain_r, float gain_g, float gain_b) {
     if (!pixels) return;
-    int total = width * height;
+    size_t total = static_cast<size_t>(width) * height;
 
     // Precompute gamma exponents
     float power_r = 1.0f / std::max(0.001f, gamma_r);
     float power_g = 1.0f / std::max(0.001f, gamma_g);
     float power_b = 1.0f / std::max(0.001f, gamma_b);
 
-    for (int i = 0; i < total; ++i) {
-        int idx = i * 3;
+    for (size_t i = 0; i < total; ++i) {
+        size_t idx = i * 3;
         float r = std::max(0.0f, pixels[idx]);
         float g = std::max(0.0f, pixels[idx + 1]);
         float b = std::max(0.0f, pixels[idx + 2]);
@@ -43,14 +43,14 @@ void ColorWheelOperator::apply_rgba(float* pixels, int width, int height,
                                      float gamma_r, float gamma_g, float gamma_b,
                                      float gain_r, float gain_g, float gain_b) {
     if (!pixels) return;
-    int total = width * height;
+    size_t total = static_cast<size_t>(width) * height;
 
     float power_r = 1.0f / std::max(0.001f, gamma_r);
     float power_g = 1.0f / std::max(0.001f, gamma_g);
     float power_b = 1.0f / std::max(0.001f, gamma_b);
 
-    for (int i = 0; i < total; ++i) {
-        int idx = i * 4;
+    for (size_t i = 0; i < total; ++i) {
+        size_t idx = i * 4;
         float r = std::max(0.0f, pixels[idx]);
         float g = std::max(0.0f, pixels[idx + 1]);
         float b = std::max(0.0f, pixels[idx + 2]);
@@ -69,10 +69,10 @@ void ColorWheelOperator::apply_cdl(float* pixels, int width, int height,
                                     const float slope[3], const float offset[3],
                                     const float power[3], int channels) {
     if (!pixels || channels < 3) return;
-    int total = width * height;
+    size_t total = static_cast<size_t>(width) * height;
 
-    for (int i = 0; i < total; ++i) {
-        int idx = i * channels;
+    for (size_t i = 0; i < total; ++i) {
+        size_t idx = i * channels;
         for (int c = 0; c < 3 && c < channels; ++c) {
             float v = std::max(0.0f, pixels[idx + c]);
             float base = std::max(0.0f, v * slope[c] + offset[c]);
