+++ AlcedoAndroid/app/src/main/cpp/core/edit/operators/film_grain_op.cpp
@@ -9,9 +9,9 @@ void FilmGrainOperator::apply(std::vector<float>& pixels, int width, int height,
     std::mt19937 gen(42);
     std::normal_distribution<float> dist(0.0f, intensity * 0.05f);
 
-    int total = width * height;
+    size_t total = static_cast<size_t>(width) * height;
     int color_channels = std::min(channels, 3); // Only apply grain to RGB, not alpha
-    for (int i = 0; i < total; ++i) {
+    for (size_t i = 0; i < total; ++i) {
         for (int c = 0; c < color_channels; ++c) {
             float& p = pixels[i * channels + c];
             p += dist(gen);
